clientlib: Renamed cache module
[GH-25]
